net/ice/base: update code with flex array safe allocations
The Linux Kernel is now requiring flex array safe allocations from drivers and
the way that we used overlaid union structures was confusing or worse, breaking
Klocwork as well as failing upstream builds that were checking -Warray-bounds.

This update was inspired by Alexander Lobakin's upstream changes to do the same,
but none of the code was copied, and the changes here were all made by me.

The ice_adminq_cmd.h file changes structures due to this change and the firmware
team may not want to sync up with these changes. However, none of the data
formats change and will continue to work with firmware using the old layouts.
It's just that after this change the shared code has provably correct usage of
flexible arrays. It's possible that the firmware team no longer really needs to
sync with this file anyway.

There are no binary format changes, and I noticed that a very similar change to
this was made and then reverted back in 2017. However, that was before we had
the STRUCT_HACK_LEN defines figured out. Now, this code compiles as well under
C++ as it does under C, and all the structure length checking macros work in all
cases that I can tell.

Signed-off-by: Jesse Brandeburg <jesse.brandeburg@intel.com>
Signed-off-by: Ian Stokes <ian.stokes@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>