net/ice/base: implement switch recipe reuse feature
If FW supports the corresponding functionality, the driver allows PFs to
subscribe the same switch recipes. Then when the PF is done with a switch
recipe, the PF can ask the FW to free that switch recipe.

When the FW detects that all subscribing PFs have freed the switch recipe, the
FW will free the switch recipe so that it can be reused.

This feature also fixes a problem where all switch recipes would eventually be
exhausted because switch recipes could not be freed, as freeing a shared recipe
could potentially break other PFs that were using it.

Signed-off-by: Steven Zou <steven.zou@intel.com>
Signed-off-by: Ian Stokes <ian.stokes@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>