net/ice/base: fix incorrect size when allocating children arrays
The ice_sched_add_root_node() and ice_sched_add_node() functions have comments
to suppress Coverity warnings about a suspicious sizeof used when allocating the
children array of an struct ice_sched_node.

The size is calculated using the size of the scheduler node, which overallocates
the array by a significant amount.

Fix the code to correctly calculate the size by using *root->children and
*node->children respectively.

This saves some memory and allows us to drop the Coverity suppression comments.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>
Signed-off-by: Ian Stokes <ian.stokes@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>