net/ice/base: re-number E810-T subdevice IDs to match upstream
When upstreaming the E810-T support for Linux, during review it was pointed out
that the E810T subdevice IDs numbering was odd. The subdevice ID numbers seemed
to be arbitrary. The upstream driver re-numbered the subdevice ID values so that
they were in increasing order.

Align with upstream, and re-number the E810-T macros to match. Fix the
ice_is_e810t function to properly use the appropriate macros when determining if
a given card is E810-T.

Note that this has no functional difference as the only users of the
ICE_SUBDEV_ID_E810T macros is ice_is_e810t. The generated switch case statements
are identical before and after this change. You can confirm this using 'gcc -E'
on Linux.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>
Signed-off-by: Ian Stokes <ian.stokes@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>