net/ice/base: update strict status when assigning BW limits
In the BW configuration performed by DCF functions, the strict/WFQ and priority
field (referred to as Generic in the EAS) is not updated in the FW.  This needs
to be updated so as to not incorrectly allocate BW credits in the traffic
shaping Tx scheduler.

Call the function "ice_sched_replay_node_prio" in the configuration flow once
the node has been determined and the value of strict verified.

Signed-off-by: Dave Ertman <david.m.ertman@intel.com>
Signed-off-by: Ian Stokes <ian.stokes@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>