net/ice/base: add E830 debug dump cluster ID values
Add E830 debug dump cluster ID values, which are different than the
values for E810. Rename E810 cluster IDs to make it easier to
distinguish which values are which.Add E830 debug dump cluster ID
values, which are different than the values for E810.
Rename E810 cluster IDs to make it easier to
distinguish which values are which.

Signed-off-by: Ian Stokes <ian.stokes@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>