net/ice/base: return high address for multi-read eth56g registers
The eth56g register checks for 40b and 64b registers do not return the high
address offset. Instead it is expected that all offsets are exactly 4 bytes
apart in the expected order. The e822 implementations explicitly return the
address in the lookup function. This form is better because it allows for
the potential of registers which may not have a static offset.

Match the e822 style by refactoring these functions to return the high
address in a function parameter.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>
Signed-off-by: Ian Stokes <ian.stokes@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>