net/ice/base: add function to read Tx timestamp status register
Both e822 and eth56g based hardware have a register indicating the status
of timestamps in the Tx timestamp memory bank. This register is used to
indicate which timestamps are currently valid. This is important because
the interrupt logic for these devices assumes that software will only read
registers with the valid bit set. If software does not follow this
practice, it can potentially cause hardware to stop reporting timestamp
interrupts.

Use of the status register value also allows software to avoid
unnecessarily reading timestamps which haven't yet been captured.

Add a helper function which reads the timestamp memory status registers for
these devices.

For e810, it is not clear if the hardware has an equivalent register.
Instead, implement a stub which just reports that all timestamps are valid.
This allows the timestamp tracking logic to use the same function for all
current device variants.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>
Signed-off-by: Ian Stokes <ian.stokes@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>