net/ice/base: support E830 in Topology AQ command
In E830, regarding the Get Tx Topology AQ command, there is a change in
the way that the RD bit must be set. For E830, the RD bit must be
cleared. Setting the RD bit for the E830 device was causing the Get Tx
Topology command to fail.

Fix it to clear RD bit for E830 device.

Signed-off-by: Dan Nowlin <dan.nowlin@intel.com>
Signed-off-by: Anatoly Burakov <anatoly.burakov@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>