net/ice: update rss lut value for RSS init
Latest shared code update changes the requirements for size definition
of rss_lut which now varies by type.

Previously it was limited to 512 even if the capabalities were greater.
With the lastest shared code update for PF type this will be 2048 and
will fail to initialise if less than this.

Signed-off-by: Ian Stokes <ian.stokes@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>