net/ring: Set mbuf->port for received packets
When using ring based ethdev, mbuf->port is not set on received packets.

For applications that use the mbuf->port to identify the incoming port,
especially when eventdev RX adapter pulls the packet on a different core
and the application running on a worker core has no clue on the incoming
port. This change adds some cycles at receive, to set the port of course.

Please advise if this change is something that can be upstreamed.

Signed-off-by: Sriram Yagnaraman <sriram.yagnaraman@ericsson.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>