net/cnxk: fix stale offload flag reset
mbuf buffer is not reset on tx and hence few fields has stale data from
previous packets. Due to stale offload flags, in one of the scenarios
with OVS, VxLAN offload flag was set while packet did not have the VxLAN
header. In the OVS packet path, the flag was read and accordingly VxLAN
processing was done but as packet did not have VxLAN header it caused
segfault.

This patch resets mbuf offload flags in rx burst function.

Fixes: 46ebc0323151 ("net/cnxk: add eswitch Rx/Tx")

Signed-off-by: Amit Prakash Shukla <amitprakashs@marvell.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>