net/ice/base: improve ice_debug_cq messages
The ice_debug_cq function is called to print debug data for a control queue
descriptor in multiple places. This includes both before we send a message on a
transmit queue, after the writeback completion of a message on the transmit
queue, and when we receive a message on a receive queue.

This function does not include data about *which* control queue the message is
on, nor whether it was what we sent to the queue or what we received from the
queue.

Modify ice_debug_cq to take two extra parameters, a pointer to the control queue
and a boolean indicating if this was a response or a command. Improve the debug
messages by replacing "CQ CMD" with a string indicating which specific control
queue (based on cq->qtype) and whether this was a command sent by the PF or a
response from the queue.

This helps make the log output easier to understand and consume when debugging.

Also tidy existing comments.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>
Signed-off-by: Ian Stokes <ian.stokes@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>