net/ice/base: remove unnecessary control queue cmd_buf arrays
The driver allocates a cmd_buf array in addition to the desc_buf array. This
array stores an ice_sq_cd command details structure for each entry in the
control queue ring.

The contents of the structure are copied from the value passed in via
ice_sq_send_cmd, and include only a pointer to storage for the write back
descriptor contents.

Originally this array was intended to support asynchronous completion including
features such as a callback function. This support was never implemented. All
that exists today is needless copying and resetting of a cmd_buf array that is
otherwise functionally unused.

Since we do not plan to implement asynchronous completions, drop this
unnecessary memory and logic. This saves memory for each control queue, and
avoids the pointless copying and memset.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>
Signed-off-by: Ian Stokes <ian.stokes@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>