net/ice/base: use "err" instead of "status" in ice_ptp_hw.c
With the exception of a few functions, the upstream Linux version of
ice_ptp_hw.c switched to using "int err" early before the initial conversion of
"enum ice_status status" to "int status" was completed. This results in the
upstream version using a different variable name than the out-of-tree shared
file uses. This leads to difference tools reporting differences due to the
variable name change.

Standardize the ice_ptp_hw.c file to use "err" instead of "status" across almost
all functions. The exceptions are:

 * ice_get_pca9575_handle
 * ice_read_sma_ctrl_e810t
 * ice_write_sma_ctrl_e810t
 * ice_is_pca9575_present

These functions were upstreamed using "int status" and we can leave them alone.
The standard for the rest of this file going forward will be to use "err" as
that matches the style used throughout the rest of the file.

This should have no impact on functionality.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>
Signed-off-by: Ian Stokes <ian.stokes@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>