net/ice/base: fix check for existing switch rule
In case the rule already exists and another VSI wants to subscribe to it new VSI
list is being created and both VSIs are moved to it. Currently, the check for
already existing VSI with the same rule is done based on fdw_id.hw_vsi_id, which
applies only to LOOKUP_RX flag. Change it to vsi_handle. This is software VSI
ID, but it can be applied here, because vsi_map itself is also based on it.

Additionally change return status in case the VSI already exists in the VSI map
to "Already exists". Such case should be handled by the caller.

Signed-off-by: Mateusz Pacuszka <mateuszx.pacuszka@intel.com>
Signed-off-by: Ian Stokes <ian.stokes@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>