net/ice/base: support DCF query port ETS adminq
In the adminq command query port ETS function, the root node teid is
needed. However, for DCF, the root node is not initialized at parent HW
init stage, which will cause error when we refer to the variable. In
this patch, we will check whether the root node is available or not first.
If not available in DCF case, after receiving virtchnl message, PF will
add the root node info before sending adminq command to FW.

Signed-off-by: Ting Xu <ting.xu@intel.com>
Signed-off-by: Ian Stokes <ian.stokes@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>