net/ice/base: misc header file clean up
Add last remaining changes that include defines, structs etc to allign with
shared code snapshot.

Add a new function called ice_fls() to the shared code.

The unsigned int ice_fls(u64 n) function finds the Most Significant Bit (MSB)
and returns the 1-based index of the bit that was set.

This means if no bits are set it returns 0. If the largest bit set is the last
bit in a 64 bit unsigned value, it will return 64.

Signed-off-by: Jesse Brandeburg <jesse.brandeburg@intel.com>
Signed-off-by: Ian Stokes <ian.stokes@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>