checkpatches: verify in-reply-to header when possible
When using checkpatches.sh locally, verify that there is an In-Reply-To
header when the patch is a respin (i.e. v2, v3, etc.). This is currently
only enforced by the upstream CI but cannot be verified locally.

This cannot be verified when checking commit ids since --in-reply-to is
a git-format-patch option which is not specified by checkpatches.sh when
generating temporary files.

Here is an example:

 $ git format-patch -v6 -1 --stdout | devtools/checkpatches.sh
 warning: [PATCH v6] graph: expose node context as pointers
 warning: respins must be --in-reply-to=<v1.patch@message.id>.
 0/1 valid patch

 $ git format-patch -v6 -1 --stdout --in-reply-to=foo | \
                     devtools/checkpatches.sh
 1/1 valid patch

Link: https://git.dpdk.org/tools/dpdk-ci/commit/?id=070b31649e48460b3
Signed-off-by: Robin Jarry <rjarry@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>