fbarray: fix incorrect lookahead behavior
Currently, whenever last bit of current index mask is set (meaning, there is
potentially a run starting at the end of the mask), lookahead loop is entered.
In that loop, if the first bit of lookahead mask is not set, the lookahead is
stopped, and the current lookahead mask index is assigned to current index mask.
However, because at that point we are inside a for-loop that increments current
index mask after each iteration, this results in erroneous mask index
increment.

Fixlookahead to avoid erroneous increment, and add corresponding unit test.

Fixes: c44d09811b40 ("eal: add shared indexed file-backed array")
Cc: stable@dpdk.org

Signed-off-by: Vipin P R <vipinp@vmware.com>
Signed-off-by: Anatoly Burakov <anatoly.burakov@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>