fbarray: fix incorrect lookbehind behavior
Currently, whenever first bit of current index mask is set (meaning, there is
potentially a run starting at the start of the mask), lookbehind loop is
entered. In that loop, if the last bit of lookbehind mask is not set, the
lookbehind is stopped, and the current lookbehind mask index is assigned to
current index mask. However, because at that point we are inside a while-loop
that decrements current index mask after each iteration, this results in
erroneous mask index decrement.

Fix lookbehind to avoid erroneous decrement, and add corresponding unit test.

Fixes: e1ca5dc86226 ("fbarray: add reverse finding of chunk")
Cc: stable@dpdk.org

Signed-off-by: Vipin P R <vipinp@vmware.com>
Signed-off-by: Anatoly Burakov <anatoly.burakov@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>