fbarray: fix lookahead ignore mask handling
When lookahead mask does not have its first bit set, we can infer that we've
lost our run. However, currently, we set ignore mask to ignore `need` number of
bits, which is incorrect because while there is no *current* run within those
bits, we might still be able to start a new run within those ignored bits later.

This issue is fixed by counting how many shifts it took to lose the run, and
this is the number of bits we should ignore (+1 to skip one we stopped on).
Also, add unit tests to reproduce the problem.

Fixes: c44d09811b40 ("eal: add shared indexed file-backed array")
Cc: stable@dpdk.org

Signed-off-by: Vipin P R <vipinp@vmware.com>
Signed-off-by: Anatoly Burakov <anatoly.burakov@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>