fbarray: fix lookbehind ignore mask handling
When lookahead mask does not have its last bit set, we can infer that we've lost
our run. However, currently, we set ignore mask to ignore first `need` bits,
which is incorrect for two reasons: first, using `need` bits as ignore bit count
means we might miss opportunities to start a new run within those bits, and more
improtantly when doing lookbehind, we start looking from the top, so we should
be ignoring *last* N bits, not *first* N bits of the mask.

This issue is fixed by counting how many shifts it took to lose the run, and
this is the number of bits we should ignore from the top (+1 to skip one we
stopped on). Also, add unit tests to reproduce the problem.

Fixes: e1ca5dc86226 ("fbarray: add reverse finding of chunk")
Cc: stable@dpdk.org

Signed-off-by: Vipin P R <vipinp@vmware.com>
Signed-off-by: Anatoly Burakov <anatoly.burakov@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>