ci: run more checks in private repositories
Though devtools/checkpatches.sh is run as part of our CI, some other
(not well known) checks could help when run in private repositories
before submitting to the mailing list and even when run from the
ovsrobot.

Most of them require a git history or checked sources to run.
And I can't guarantee there won't be false positives.

Add a new job just for those checks so that it won't block compilation
tests in other jobs.

Signed-off-by: David Marchand <david.marchand@redhat.com>
Signed-off-by: Robin Jarry <rjarry@redhat.com>
Acked-by: Aaron Conole <aconole@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>